requirements.txt
plotly.express
dash
pandas